Use monolithic tarball to build
This seriously simplifies the build process and reduces the amount of gynmastics we need to moving all the tarballs into the right locations. As a result the package.yml is much easier to read at a glance.

As a bonus, we can now easily enable or disable various llvm sub-projects via the _PROJECTS env for x86_64 and emul32.

**Packaging**
- Remove an unneccessary LLD patch with this build method, this stops the compact_unwind_encoding.h header from being installed.
- Remove unncecessary update_sources.sh script